requirements.txt
numpy
pandas
tensorflow-gpu
keras
h5py
pybel